adjustments

+++ src/asammdf/blocks/utils.py
@@ -820,7 +820,7 @@ def as_non_byte_sized_signed_int(integer_array: NDArray[Any], bit_length: int) -
         truncated_integers = integer_array & ((1 << bit_length) - 1)  # Zero out the unwanted bits
     return where(
         truncated_integers >> bit_length - 1,  # sign bit as a truth series (True when negative)
-        (2**bit_length - truncated_integers) * -1,  # when negative, do two's complement
+        (2**bit_length - truncated_integers) * np.int8(-1),  # when negative, do two's complement
         truncated_integers,  # when positive, return the truncated int
     )
